init
Start of recreation begins! A new personal page to replace my current conorhennessy.uk page. This is made from scratch to actually have something on the page - rather than an 'in progress' page.

The page features a random contour effect to the left and an animated feature name to the top. Below features an intro, socials and some projects.